refactor(#21): ShopItemResponse dto 파일 이동
