+++ src/main/java/software_capstone/backend/app/store/controller/ShopController.java
@@ -8,7 +8,7 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 import software_capstone.backend.app.store.document.category.ItemCategory;
-import software_capstone.backend.app.store.dto.ShopItemResponse;
+import software_capstone.backend.app.store.dto.response.ShopItemResponse;
 import software_capstone.backend.app.store.service.ShopService;
 
 import java.util.List;
